warehouse_system/backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-filter
drf-spectacular
python-decouple
Pillow
django-cors-headers
gunicorn